Made the Color Tab Modular
The Object that the Color Tab relies on was also made modular. That Object is the `tabStates` object which further relies on the `util` object, which was also made modular.
